requirements.txt
asgiref==3.7.2
certifi==2023.7.22
chardet==4.0.0
Django==4.2.6
djangorestframework==3.14.0
idna==2.10
Pillow==9.0.1
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.25.1
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==1.26.17
